requirements.txt
Django==3.0.5
django-filter==2.2.0
django-timedeltatemplatefilter==0.1.2
openpyxl==3.0.3
